grpc-sentiment-service/requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
redis==5.0.1
protobuf==4.24.4
psycopg2-binary==2.9.9 
